feat: 기본 폴더 생성 로직 InventoryFolderService로 이동
