+++ backend/src/main/java/capstone/backend/global/security/oauth2/service/CustomOAuth2UserService.java
@@ -1,9 +1,9 @@
 package capstone.backend.global.security.oauth2.service;
 
+import capstone.backend.domain.inventory.service.InventoryFolderService;
 import capstone.backend.domain.member.repository.MemberRepository;
 import capstone.backend.domain.member.scheme.Member;
 import capstone.backend.domain.member.scheme.Role;
-import capstone.backend.domain.member.service.MemberService;
 import capstone.backend.global.security.oauth2.user.CustomOAuth2User;
 import capstone.backend.global.security.oauth2.user.OAuth2UserInfoFactory;
 import capstone.backend.global.security.oauth2.user.OAuth2UserInfo;
@@ -25,7 +25,7 @@ import org.springframework.stereotype.Service;
 public class CustomOAuth2UserService extends DefaultOAuth2UserService {
 
     private final MemberRepository memberRepository;
-    private final MemberService memberService;
+    private final InventoryFolderService inventoryFolderService;
 
     @Override
     public OAuth2User loadUser(OAuth2UserRequest oAuth2UserRequest) throws OAuth2AuthenticationException {
@@ -69,7 +69,7 @@ public class CustomOAuth2UserService extends DefaultOAuth2UserService {
                         memberRepository.save(newMember);
 
                         //기본 폴더 생성
-                        memberService.createDefaultFolder(newMember);
+                        inventoryFolderService.createDefaultFolder(newMember);
 
                         return newMember;
                     });
